Move server.py to support/server.py
